Update logic.ps1
Moving away from params as these seem to at this time be failing and I think in the end I'm better off with registry to track everything I need to pass between instances. Also #Requires was removed as I don't want to have to run the sript a special way just to elevate. This is abut simplicity